Naive configs grouping & flatter hierarchy
As I started extracting homebrew, I've realised 2 things:

1. I need a cleanup script which removes homebrew's global state

After the first run, order of config execution no longer matters. On a
clean system, the setup script will break. I see no better alternative
than:

2. Define config execution order

In the olden days, 00_exec_1, 10_exec_2 was a thing. I don't want to be
that crude and brittle, but I will need to borrow some of those ideas
for configs such as homebrew which needs to be installed & taps need to
be applied before the rest of the config can be run. It will be painful,
but I can't see a better alternative.