Automatically download Ruby Tapas
Leverages ruby-tapas-downloader Ruby gem. Favours, rbenv but does not
enforce it.

Ruby Tapas login credentials & download config is stored in the defualt
Keychain. An iCloud Keychain won't work as it's not integrated with the
security utility. Not ideal, but still better than storing these
credentials in plaintext.

I briefly considered [Vault](https://www.vaultproject.io/) but concluded
that the extra setup required was not worth the effort. Keychain is
built into the OS and it's been working perfectly for many years. Once
Vault matures it might be a better solution, but for now, I choose the
"old, boring & reliable".