Revert "Change default shell to bash v4"
This reverts commit 12ca50c7a04ccb6093ea09bbc8fb80ffbb0d5e87.

On second thought, this seems too intrusive. I want to deviate from the
defaults as little as possible. The main reason why I wanted this is to
get the advanced features of the language such as ${var^^}, but I might
be barking the wrong tree. Golang or Rust are better suited for my
purpose, not the swiss army knife of scripting and cheap system services.