Same shell prompt as my vim status bar
Even though the shell prompt config could be generated at runtime, I
prefer to have it version controlled so that I can track any changes. If
a powerline vim plugin update will break this integration, I might
re-consider this, but for now happy to go with the naive but simple
approach.