Install YouCompleteMe every single time to avoid weird bugs
If the YouCompleteMe vim plugin gets upgraded, the install.sh
that it ships with needs to be re-run. I could figure out on how to
trigger this install only on upgrade, but it doesn't feel
straightforward or simple, so I'm settling for the naive and run it on
every setup run.