Add kubectl to homebrew formulae
Even though kubectl gets installed by the docker cask (a.k.a. Docker
Desktop), we want the latest kubectl version, so that we can make use of
all the new features that may be available in various K8S installations
that we target.

Picking up Brewfile.lock.json in this commit as well, it seems to be a
new file generated by the latest brew.